workout_app/requirements.txt
amqp==5.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.13
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
kombu==5.4.2
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
